Refactoring, fix compiler warnings
