+++ task2/server/message.h
@@ -1,0 +1,13 @@
+//
+// Created by Malik Hiraev on 15.11.2019.
+//
+
+#ifndef NETWORKSLAB2019_MESSAGE_H
+#define NETWORKSLAB2019_MESSAGE_H
+
+typedef struct Message {
+    char *msg; // ссылка на сообщеньку
+    size_t size; // размер сообщения
+} Message;
+
+#endif //NETWORKSLAB2019_MESSAGE_H
